pull struct TestSet to a new header
