+++ src/test/test_set.hpp
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <memory>
 #include <string>
 #include <vector>
 
@@ -41,39 +40,6 @@ struct NormalTestSet {
 	std::vector<TestFunction> m_testers;
 
 	TestReport execute();
-};
-
-struct TestSet {
-	struct ITestSet {
-		virtual TestReport execute() = 0;
-		virtual ~ITestSet() = default;
-	};
-
-	template <typename T>
-	struct TestSetImpl : ITestSet {
-		T x;
-
-		TestSetImpl(T x_)
-		    : x {std::move(x_)} {}
-
-		TestReport execute() override {
-			return x.execute();
-		}
-	};
-
-	std::unique_ptr<ITestSet> m_data;
-
-	template <typename T>
-	TestSet(T data)
-	    : m_data {std::make_unique<TestSetImpl<T>>(std::move(data))} {}
-
-	bool operator==(TestSet const& o) const {
-		return m_data == o.m_data;
-	}
-
-	TestReport execute() {
-		return m_data->execute();
-	}
 };
 
 } // namespace Test
